requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
Django==5.1.5
flatbuffers==24.12.23
humanfriendly==10.0
idna==3.10
imageio==2.36.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.20.1
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
PyMatting==1.1.13
pyreadline3==3.5.4
referencing==0.36.1
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.25.0
scipy==1.15.1
sqlparse==0.5.3
sympy==1.13.3
tifffile==2025.1.10
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
